+++ CHANGELOG.md
@@ -15,6 +15,7 @@ to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 - Ability to decompose meshes into individual line segments (and back).
     - Break meshes into segments with Flynn._3DMesh.to_segments()
     - Make meshes from segments with Flynn._3DMeshFromSegments()
+- screenOffsetV argument to Flynn._3DRenderer.
 
 ## 3.19.1 - 2020-12-12
 ### Added